+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Updated Pow requirement to `~> 1.0.19`.
 
+### Enhancements
+
+* [`PowAssent.Plug`] `PowAssent.Plug.callback/4` now adds the `userinfo` with claims to the user identity params
+
 ## v0.4.6 (2020-02-16)
 
 The callback flow has been changed so sessions are now stored in the backend cache with `PowAssent.Store.SessionCache` instead of using `Plug.Session`. This prevents exposure of sensitive data, as the only thing stored in the Plug session is a random UUID.